+++ tests/factories.py
@@ -17,6 +17,7 @@ class EncryptedModelFactory(factory.DjangoModelFactory):
     integer_pgp_pub_field = 42
     pgp_pub_field = factory.Sequence('Text with public key {}'.format)
     pgp_pub_date_field = datetime.date(year=2000, month=1, day=1)
+    pgp_pub_null_boolean_field = True
 
     email_pgp_sym_field = factory.Sequence('email{}@symmetric.key'.format)
     integer_pgp_sym_field = 43