+++ tests/models.py
@@ -12,6 +12,7 @@ class EncryptedModel(models.Model):
     integer_pgp_pub_field = fields.IntegerPGPPublicKeyField(blank=True, null=True)
     pgp_pub_field = fields.TextPGPPublicKeyField(blank=True, null=True)
     pgp_pub_date_field = fields.DatePGPPublicKeyField(blank=True, null=True)
+    pgp_pub_null_boolean_field = fields.NullBooleanPGPPublicKeyField()
 
     email_pgp_sym_field = fields.EmailPGPSymmetricKeyField(blank=True, null=True)
     integer_pgp_sym_field = fields.IntegerPGPSymmetricKeyField(blank=True, null=True)